Refactor: move resolve_module_placeholders from constants.py to utils.py
Constants should only contain constant values, not functions with logic.
The helper function now lives in dcim/utils.py alongside other utilities
like update_interface_bridges and create_port_mappings.
